+++ dbt/adapters/vertica/__init__.py
@@ -16,6 +16,7 @@
 from dbt.adapters.vertica.connections import verticaConnectionManager
 from dbt.adapters.vertica.connections import verticaCredentials
 from dbt.adapters.vertica.impl import verticaAdapter
+from dbt.adapters.vertica.column import VerticaColumn
 
 from dbt.adapters.base import AdapterPlugin
 from dbt.include import vertica