+++ dbt/include/vertica/macros/materializations/models/incremental/merge.sql
@@ -1,21 +1,23 @@
 {% macro vertica__get_merge_sql(target_relation, tmp_relation, unique_key, dest_columns) %}
   {%- set dest_columns_csv =  get_quoted_csv(dest_columns | map(attribute="name")) -%}
   {%- set merge_columns = config.get("unique_key", default=None)%}
-  {%- set merge_update_columns = config.get("merge_update_columns", default=dest_columns)-%}
-  {% if merge_update_columns %}
-    {%- set merge_update_columns = dest_columns -%}
-  {% endif %}
-
+  {%- set merge_update_columns = config.get("merge_update_columns")-%}
+ 
   merge into {{ target_relation }} as DBT_INTERNAL_DEST
   using {{ tmp_relation }} as DBT_INTERNAL_SOURCE
-
+  
+ 
   {#-- Test 1, find the provided merge columns #}
   {% if merge_columns %}
-    on 
-    {% for column in [merge_columns] %}
+    on
+    {% if merge_columns is string %}
+      DBT_INTERNAL_DEST.{{ adapter.quote(merge_columns) }} = DBT_INTERNAL_SOURCE.{{ adapter.quote(merge_columns) }}
+    {% else %}
+    {% for column in merge_columns -%}
       DBT_INTERNAL_DEST.{{ adapter.quote(column) }} = DBT_INTERNAL_SOURCE.{{ adapter.quote(column) }}
       {%- if not loop.last %} AND {% endif %} 
     {%- endfor %}
+    {% endif %}
   {#-- Test 2, use all columns in the destination table #}
   {% else %}
     on
@@ -26,10 +28,20 @@
   {% endif %}
 
   when matched then update set
-  {% for column in merge_update_columns -%}
-    {{ adapter.quote(column.name) }} = DBT_INTERNAL_SOURCE.{{ adapter.quote(column.name) }}
-    {%- if not loop.last %}, {% endif %}
-  {%- endfor %}
+  {% if merge_update_columns %}
+    
+    {% for column in merge_update_columns -%}
+      {{ adapter.quote(column) }} = DBT_INTERNAL_SOURCE.{{ adapter.quote(column) }}
+      {%- if not loop.last %}, {% endif %}
+    {%- endfor %}
+  {% else %}
+    
+    {% for column in dest_columns -%}
+      {%- set merge_update_columns = dest_columns -%}
+      {{ adapter.quote(column.name) }} = DBT_INTERNAL_SOURCE.{{ adapter.quote(column.name) }}
+      {%- if not loop.last %}, {% endif %}
+    {%- endfor %}
+  {% endif %}
 
   when not matched then insert
     ({{ dest_columns_csv }})
@@ -47,12 +59,29 @@
     {%- set dest_cols_csv = get_quoted_csv(dest_columns | map(attribute="name")) -%}
 
     {% if unique_key %}
+      {% if unique_key is string %}
         delete from {{ target }}
-            where (
-                {{ unique_key }}) in (
-                select ({{ unique_key }})
-                from {{ source }}
-            );
+            where 
+                ({{ unique_key }}) in (
+                  select ({{ unique_key }})
+                  from {{ source }}
+                )
+              
+                
+            ;
+      {% else %}
+        delete from {{ target }}
+            where 
+              {% for column in unique_key -%}
+                ({{ column }}) in (
+                  select ({{ column }})
+                  from {{ source }}
+                )
+                {%- if not loop.last %} AND {% endif %} 
+              {%- endfor %}
+                
+            ;
+      {% endif%}
 
     {% endif %}