+++ dbt/include/vertica/macros/materializations/models/table/create_table_as.sql
@@ -9,6 +9,11 @@
   {%- set partition_by_string = config.get('partition_by_string', default=none) -%}
   {%- set partition_by_group_by_string = config.get('partition_by_group_by_string', default=none) -%}
   {%- set partition_by_active_count = config.get('partition_by_active_count', default=none) -%}
+  {%- set contract_config = config.get('contract') -%}
+  
+  {% if contract_config.enforced %}
+     {{exceptions.warn("Model contracts cannot be enforced by <adapter>!")}}
+  {% endif %}
   
   create {% if temporary: -%}local temporary{%- endif %} table
     {{ relation.include(database=(not temporary), schema=(not temporary)) }}
@@ -16,11 +21,12 @@
     INCLUDE SCHEMA PRIVILEGES as (
     {{ sql }}
   )
+ {% if not temporary: %}
 
   {% if order_by is not none  -%}
       order by {{ order_by }} 
   {% endif -%}
-
+  
   {% if segmented_by_string is not none -%}
               segmented  BY  {{ segmented_by_string }} {% if segmented_by_all_nodes %} ALL NODES {% endif %}
   {% endif %}
@@ -42,6 +48,7 @@
       SET ACTIVEPARTITIONCOUNT {{ partition_by_active_count }}
     {% endif %}
   {% endif %}  
+ {% endif %}
   ;
 {% endmacro %}